Add meta gather audit pair
