+++ audits/viewport-meta-tag/audit.js
@@ -1,0 +1,37 @@
+/**
+ * @license
+ * Copyright 2015 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+function hasCorrectViewportMetaQuery(obj) {
+  if (obj.type === "object" && obj.subtype === "null")
+    ret = { pass : false };
+  else if (obj.subtype === 'node' && obj.props.content.includes('width='))
+    ret = { pass: true };
+  else
+    throw new Error("Unexpected viewport elements.");
+  return ret;
+};
+
+
+module.exports = function(data) {
+  if (data.viewportMetaQuery === undefined) {
+    throw new Error('viewport meta auditor requires viewport meta query data');
+  }
+
+  return Promise.resolve({
+    hasCorrectViewportMetaQuery: hasCorrectViewportMetaQuery(data.viewportMetaQuery),
+  });
+}
